chore: move resume template into /template folder
