+++ src/app/resume-editor/components/template/styles.ts
